Rename time chunking functions
