+++ public/app/plugins/datasource/loki/logsTimeChunking.test.ts
@@ -1,6 +1,6 @@
-import { getRangeChunks } from './logsTimeSplit';
+import { getRangeChunks } from './logsTimeChunking';
 
-describe('querySplit', () => {
+describe('logs getRangeChunks', () => {
   it('should split time range into chunks', () => {
     const start = Date.parse('2022-02-06T14:10:03.234');
     const end = Date.parse('2022-02-06T14:11:03.567');
