+++ public/app/plugins/datasource/loki/metricTimeChunking.test.ts
@@ -1,6 +1,6 @@
-import { getRangeChunks } from './metricTimeSplit';
+import { getRangeChunks } from './metricTimeChunking';
 
-describe('querySplit', () => {
+describe('metric getRangeChunks', () => {
   it('should split time range into chunks', () => {
     const start = Date.parse('2022-02-06T14:10:03');
     const end = Date.parse('2022-02-06T14:11:03');
